+++ nvim/after/plugin/conform.lua
@@ -1,10 +1,34 @@
+--  I'd prefer not to write this twice but imports from local lsp.lua weren't working for me
+local lspconfig = require('lspconfig')
+local capabilities = require('cmp_nvim_lsp').default_capabilities()
+
+local M = {}
+
+function M.organize_imports()
+  local params = {
+    command = "_typescript.organizeImports",
+    arguments = { vim.api.nvim_buf_get_name(0) },
+    title = ""
+  }
+  vim.lsp.buf.execute_command(params)
+end
+
+lspconfig.tsserver.setup({
+  capabilities = capabilities,
+  commands = {
+    OrganizeImports = {
+      M.organize_imports,
+      description = "Organize Imports"
+    }
+  }
+})
+
 require("conform").setup({
   formatters_by_ft = {
     lua = { "stylua" },
     -- Conform will run multiple formatters sequentially
     python = { "black" },
     go = { "goimports", "gofumpt", "goimports-reviser" },
-
     -- Use a sub-list to run only the first available formatter
     javascript = { { "prettier", "eslint_d", "biome", "biome-check" } },
     css = { { "prettier", "biome", "biome-check" } },
@@ -41,6 +65,7 @@ require("conform").setup({
 vim.api.nvim_create_autocmd("BufWritePre", {
   pattern = "*",
   callback = function(args)
+    M.organize_imports()
     require("conform").format({ bufnr = args.buf })
   end,
 })
@@ -63,6 +88,7 @@ vim.api.nvim_create_user_command("Format", function(args)
     }
   end
   require("conform").format({ async = true, lsp_fallback = true, range = range })
+  M.organize_imports()
 end, { range = true })
 
 vim.api.nvim_set_keymap('n', '<leader>f', ':Format<CR>', { silent = true })
